addRemove test
